refactor(logger): move stringifyBigInts to @prosopo/util
