+++ packages/logger/package.json
@@ -28,6 +28,7 @@
 	"author": "Prosopo Limited",
 	"license": "Apache-2.0",
 	"dependencies": {
+		"@prosopo/util": "3.2.14",
 		"zod": "3.23.8"
 	},
 	"devDependencies": {
