+++ packages/logger/src/logger.ts
@@ -12,7 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { stringifyBigInts } from "@prosopo/util";
 import { z } from "zod";
+
+export { stringifyBigInts } from "@prosopo/util";
 
 export type LogObject = object;
 export type LogRecord = {
@@ -348,35 +351,3 @@ export class NativeLogger implements Logger {
 		}
 	}
 }
-
-/**
- * Recursively converts BigInt values to strings throughout an object, array, or primitive value.
- *
- * BigInts must be cast to strings before applying JSON.stringify(), as it cannot serialize BigInt values and will
- * throw "TypeError: Do not know how to serialize a BigInt".
- *
- * @param value - The value to process (can be a primitive, object, or array)
- * @returns The same value with all BigInt instances converted to strings
- */
-export const stringifyBigInts = (value: unknown): unknown => {
-	if ("bigint" === typeof value) {
-		return value.toString();
-	}
-
-	if (isObject(value)) {
-		for (const key of Object.keys(value)) {
-			value[key] = stringifyBigInts(value[key]);
-		}
-	}
-
-	if (Array.isArray(value)) {
-		for (let i = 0; i < value.length; i++) {
-			value[i] = stringifyBigInts(value[i]);
-		}
-	}
-
-	return value;
-};
-
-const isObject = (value: unknown): value is Record<string, unknown> =>
-	Object === value?.constructor;
